Fix AM_CFLAGS placement and adjust configure.ac to prevent unwanted flags
The AM_CFLAGS definition is on the wrong place, it should be on
src/Makefile.am, instead of Makefile.am. Also configure.ac is adding
some flags on it's own.

To solve this, move the AM_CFLGAS to the correct place and add make
CFLAGS empty before setting AC_PROG_CC on configure.ac. After that,
remove all warnings in the code.