vinumc: Start to recognize '$*' as a reference to all args
The language has a way to reference all the args in a function
definition with '$*'.

As we can only decide if a '$*' is a arg reference by the location of
the token, the lexer returns the token `WORD` but it create a node with
the correct token `ARG_REF_ALL_ARGS`.