src/ast: Create a function to print the AST in the dot language
I think it would be useful for visualising the AST if we could print it
in the dot language[1]. With this we could use the output with the `dot`
utility to generate the visualisation.

To do this, create a new function `ast_dot()` that prints the AST to a
`FILE*` in the dot language.

[1]: https://en.wikipedia.org/wiki/DOT_(graph_description_language)