Fixing chained calls error
This error was resolved with changes to the parser. A new "call" rule has been created, which takes care of function calls in the ".function" format. Furthermore, the construction of "program" is given by a more complex string manipulation instead of just concatenations, given that the "call" rule returns to "program" with the prepared block, simply inserting its content in the correct position.

NOTE: the characters '(', ')' and '.' are not yet allowed as block content.