src/vinumc.c: Change output stream to stdout
Currently, we write the evaluation output to adhoc "out.txt" file, as we
don't have the means to let the user chose the output file.

It's better to just have the output written the output to stdout, as
this gives more freedom to the user via redirection, and also doesn't
clobber the project files.