src/eval.c: Create the needed scopes if a symbol definition has calls
When a symbol `X` has a definition which contains calls to other
symbols, those calls need resolution when the symbol `X` is
called.

For those inside calls to be resolved they need their own scopes, to
create them, run `resolve_symbol` on the new nodes that are created when
calling a symbol.