src/dry_*: Handle args ref only with flex and bison
I was sleepy and dumb wen I wrote the first ARG_REF_ALL_ARGS
implementation. It's a lot simpler to just treat all the logic on flex
and bison.

Also, this implementation is buggy, if there is no `WORD` with the
`ARG_REF_ALL_ARGS` it will leave a `TEXT` node with not children behind.