Separate each project into its own directory
The project is growing, now we have two projects (vinumc and vunit) in
the future we're going to have more. To organize this, turn this repo
into a monorepo[1], were all vinum related projects reside.

The Autools build system is not up to the task (the SUBDIRS situation
is really painful when doing something like this).

So, change the build system to meson[2], using it's subproject
feature[3] makes the whole thing really manageable. Also, not having a
bunch of build scripts scattered around the project is a positive.

[1]: https://en.m.wikipedia.org/wiki/Monorepo 
[2]: https://mesonbuild.com/
[3]: https://mesonbuild.com/Subprojects.html

Fixes: #25
Fixes: #34
Fixes: #34