vinumc: Create an argument parsing system
This argument for loop is not future proof, is better to use the libc
getopt_long function for argument parsing.

Also, is good to have a way to declare new flags in a centralized
location, so abstract the getopt function with a `parse_cmdline`
function that receives a list of `struct flag`.